Add missing macro in pkg-config file
When compiling the library with OPT=big or OPT=LITTLE, the makefile
defines a macro that affects the layout of the armv8_cipher_constants_t
structure, which is part of the API. Include the macro in Cflags in
the generated pkg-config file so that the user of the library gets
compiled with the same layout.

This fixes an out-of-bounds array write by AES key expansion when the
library has been built with one of the OPT flags and the application
using the library has been compiled using pkg-config without manually
adding the needed PERF_GCM_* macro.

Signed-off-by: Janne Peltonen <janne.peltonen@nokia.com>