Generate pkg-config file from a template
The file for pkg-config was generated line-by-line in Makefile.
This required escaping and was developer-unfriendly.
Generate the .pc file from a template file using awk
(which is already used and must be on all POSIX-compliant systems).
While there, alter the target name for the .pc file
so that it is not needlessly rebuilt.

Signed-off-by: Dmitry Kozlyuk <dmitry.kozliuk@gmail.com>