Add 'install' target
Makefile suggested that the library should only be used from the build
directory, which currently must be the same as the source directory.
Add "install" make target and handle convenient variables:

* PREFIX (default: /usr/local)
* INCLUDEDIR (relative to PREFIX, default: include)
* LIBDIR (relative to PREFIX, default: lib)

There is no need to use this target
to continue using the library from the build directory,
but it saves work for users and packagers.

Signed-off-by: Dmitry Kozlyuk <dmitry.kozliuk@gmail.com>