Remove undefined behavior from reference implementation
The C+intrinsics reference implementation performed a shift >= 64
bits in the AES-GCM decrypt functions, this causes UB and may cause
the test (e.g. using NIST test vectors) to fail.
Bitwise-and the shift amount with 63 ("zero_bits & 63") to make
the shift defined again.
Modify the corresponding code in the AES-GCM encrypt functions to
use the same bitwise-and for consistency (used a subtraction instead).

Note: The reference implementation is not built by default (requires
changes to the Makefile) and is not an official delivery of the
AArch64cryptolib repository.

Bug reported by stephengao0121:
https://github.com/ARM-software/AArch64cryptolib/issues/7

Change-Id: I1d4042512071ffe2b5b36ad90bf2625dffc207be
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/c/attk/AArch64cryptolib/+/766850
Tested-by: Ola Liljedahl <ola.liljedahl@arm.com>
Reviewed-by: Ola Liljedahl <ola.liljedahl@arm.com>